Move username to dropdown menu on right side
Replaced the left-side username display with a dropdown menu positioned on the right side of the header. On desktop, the menu shows the username with a dropdown arrow. On mobile, it displays as a hamburger menu icon. The dropdown contains the logout option and is designed to accommodate additional menu items in the future.